+++ src/misc/assert.ts
@@ -1,5 +1,3 @@
 export function assert(v: any, msg?: string): asserts v {
-    if (!v) {
-        throw new Error(`Assertion error${msg ? `: ${msg}` : ''}`)
-    }
+    if (!v) throw new Error(`Assertion error${msg ? `: ${msg}` : ''}`)
 }