+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - Fix ig.ENTITY.EnemyCounter ding on remote on second remote client creation
 - Fix ig.ENTITY.Effect not spawning on remote after a while
 - Fix crash when loading maps in areas with more than one floor
+- Fix creating server from current state resetting stuff like vars, active landmarks etc.
+- Fix saving server state not saving stuff like vars, active landmarks etc.
 
 ## [0.5.7] 2025-11-10